20-contrato-endpoint-de-retorno-das-informações-de-avaliações-para-exibir-o-gráfico - alterações endpoint do gráfico
