+++ gastos/src/main/java/br/com/publico/gastos/repository/ColaboradorRepository.java
@@ -1,10 +1,8 @@
 package br.com.publico.gastos.repository;
 
-import br.com.publico.gastos.domain.model.Avaliacao;
 import br.com.publico.gastos.domain.dto.response.GraficoAvaliacoesResponse;
 import br.com.publico.gastos.domain.model.Colaborador;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
@@ -21,9 +19,8 @@ public interface ColaboradorRepository extends JpaRepository<Colaborador, Long> 
     @Query("select c from Colaborador c where c.sigla in :siglas")
     List<Colaborador> findBySigla(List<String> siglas);
 
-    @Query("SELECT new br.com.publico.gastos.domain.dto.response.GraficoAvaliacoesResponse(c.id, c.nome, c.sigla, c.avaliacoes) " +
+    @Query("SELECT new br.com.publico.gastos.domain.dto.response.GraficoAvaliacoesResponse(c.id, c.nome, c.sigla) " +
            "FROM Colaborador c " +
-           "LEFT JOIN Avaliacao a ON a.colaborador.id = c.id " +
            "WHERE c.id in :id ")
     List<GraficoAvaliacoesResponse> findByIdIn(@Param("id") List<Long> id);
 }
