+++ gastos/src/main/java/br/com/publico/gastos/services/AvaliacaoService.java
@@ -1,11 +1,11 @@
 package br.com.publico.gastos.services;
 
 
-import br.com.publico.gastos.domain.model.Avaliacao;
-import org.springframework.web.multipart.MultipartFile;
 import br.com.publico.gastos.controller.request.AvaliacaoRequest;
 import br.com.publico.gastos.controller.request.AvaliacaoUpdateRequest;
 import br.com.publico.gastos.domain.dto.response.AvaliacaoResponse;
+import br.com.publico.gastos.domain.model.Avaliacao;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
 
@@ -23,4 +23,6 @@ public interface AvaliacaoService {
     List<AvaliacaoResponse> buscarAvaliacoes();
 
     void deletar(Long avaliacaoId);
+
+    List<Avaliacao> buscarAvaliacoesPorIdColaborador(Long id);
 }
