+++ gastos/src/main/java/br/com/publico/gastos/services/impl/AvaliacaoServiceImpl.java
@@ -6,10 +6,15 @@ import br.com.publico.gastos.controller.request.AvaliacaoRequest;
 import br.com.publico.gastos.controller.request.AvaliacaoUpdateRequest;
 import br.com.publico.gastos.domain.dto.mapper.AvaliacaoMapper;
 import br.com.publico.gastos.domain.dto.response.AvaliacaoResponse;
-import br.com.publico.gastos.domain.model.*;
+import br.com.publico.gastos.domain.model.Avaliacao;
+import br.com.publico.gastos.domain.model.Colaborador;
+import br.com.publico.gastos.domain.model.Status;
+import br.com.publico.gastos.domain.model.TipoAvaliacao;
+import br.com.publico.gastos.domain.model.TipoResultado;
 import br.com.publico.gastos.repository.AvaliacaoRepository;
 import br.com.publico.gastos.repository.ColaboradorRepository;
 import br.com.publico.gastos.services.AvaliacaoService;
+import br.com.publico.gastos.services.exception.ColaboradorPossuiAvaliacaoException;
 import br.com.publico.gastos.services.exception.EntidadeNaoEncontradaException;
 import br.com.publico.gastos.services.message.ValidationMessage;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,21 +23,12 @@ import org.springframework.web.multipart.MultipartFile;
 
 import javax.transaction.Transactional;
 import java.io.IOException;
-import java.util.List;
-
-import br.com.publico.gastos.controller.request.AvaliacaoRequest;
-import br.com.publico.gastos.controller.request.AvaliacaoUpdateRequest;
-import br.com.publico.gastos.domain.dto.mapper.AvaliacaoMapper;
-import br.com.publico.gastos.domain.dto.response.AvaliacaoResponse;
-import br.com.publico.gastos.domain.model.Status;
-import br.com.publico.gastos.domain.model.TipoAvaliacao;
-import br.com.publico.gastos.domain.model.TipoResultado;
-import br.com.publico.gastos.services.exception.ColaboradorPossuiAvaliacaoException;
-
-import javax.transaction.Transactional;
 import java.time.LocalDate;
 import java.time.temporal.TemporalAdjusters;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -173,5 +169,10 @@ public class AvaliacaoServiceImpl implements AvaliacaoService {
         }
         var avaliacao = avaliacaoOptional.get();
         avaliacaoRepository.delete(avaliacao);
+    }
+
+    @Override
+    public List<Avaliacao> buscarAvaliacoesPorIdColaborador(Long id) {
+        return avaliacaoRepository.findByColaboradorId(id);
     }
 }
