+++ gastos/src/main/java/br/com/publico/gastos/services/impl/ColaboradorServiceImpl.java
@@ -1,9 +1,12 @@
 package br.com.publico.gastos.services.impl;
 
 import br.com.publico.gastos.controller.request.ColaboradorRequest;
+import br.com.publico.gastos.domain.dto.AvaliacaoDTO;
 import br.com.publico.gastos.domain.dto.mapper.ColaboradorMapper;
 import br.com.publico.gastos.domain.dto.response.GraficoAvaliacoesResponse;
+import br.com.publico.gastos.domain.model.Avaliacao;
 import br.com.publico.gastos.repository.ColaboradorRepository;
+import br.com.publico.gastos.services.AvaliacaoService;
 import br.com.publico.gastos.services.ColaboradorService;
 import br.com.publico.gastos.services.exception.ColaboradorPossuiAvaliacaoException;
 import br.com.publico.gastos.services.exception.EntidadeNaoEncontradaException;
@@ -16,7 +19,11 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
 import javax.transaction.Transactional;
+import java.time.LocalDate;
+import java.time.format.TextStyle;
 import java.util.List;
+import java.util.Locale;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -27,6 +34,9 @@ public class ColaboradorServiceImpl implements ColaboradorService {
 
     @Autowired
     private ColaboradorMapper colaboradorMapper;
+
+    @Autowired
+    private AvaliacaoService avaliacaoService;
 
     @Override
     @Transactional
@@ -93,6 +103,30 @@ public class ColaboradorServiceImpl implements ColaboradorService {
 
     @Override
     public List<GraficoAvaliacoesResponse> obterInformacoesGrafico(List<Long> idColaboradores) {
-        return repository.findByIdIn(idColaboradores);
+        List<GraficoAvaliacoesResponse> graficoAvaliacoesResponses = repository.findByIdIn(idColaboradores);
+
+        graficoAvaliacoesResponses.forEach(avaliacoesResponse -> {
+            avaliacoesResponse.setAvaliacoes(avaliacaoService.buscarAvaliacoesPorIdColaborador(avaliacoesResponse.getIdColaborador())
+                    .stream().map(this::avaliacaoEntityToDTO)
+                    .collect(Collectors.toList()));
+        });
+
+        return graficoAvaliacoesResponses;
+    }
+
+    private AvaliacaoDTO avaliacaoEntityToDTO(Avaliacao avaliacao) {
+        AvaliacaoDTO avaliacaoDTO = new AvaliacaoDTO();
+
+        avaliacaoDTO.setId(avaliacao.getId());
+        avaliacaoDTO.setMesAno(formataMesAnoAvaliacao(avaliacao.getData()));
+        avaliacaoDTO.setNota(avaliacao.getNota());
+        avaliacaoDTO.setResultado(avaliacao.getResultado().getDescricao());
+
+        return avaliacaoDTO;
+    }
+
+    private String formataMesAnoAvaliacao(LocalDate data) {
+        Locale brasil = new Locale("pt", "BR");
+        return data.getMonth().getDisplayName(TextStyle.SHORT, brasil).concat("/").concat(Integer.toString(data.getYear()));
     }
 }
